add fallible ArrayString methods
Added `try_push`, `try_push_str`, `try_insert`, and `try_insert_str`.
These methods return errors on overflow instead of panicking. The
insertion methods still panic when the provided index is out of bounds
or not on a char boundary, as this is still programmer error, and the
goal is to provide easier handling of the unique aspect of ArrayString,
which is a fixed capacity.